+++ src/station-interface/middleware/index.js
@@ -1,0 +1,7 @@
+import Auth from './auth.js'
+import Raw from './raw.js'
+
+export default Object.freeze({
+    Auth,
+    Raw,
+})
