+++ src/station-interface/middleware/raw.js
@@ -1,0 +1,5 @@
+import bodyParser from 'body-parser'
+
+export default bodyParser.raw({
+  limit: '50mb'
+})
