+++ src/station-interface/routes/blu.js
@@ -1,0 +1,3 @@
+export default Object.freeze((req, res) => {
+  res.render('main-blu', { title: 'CTT Blu Receiver Interface' })
+})
