+++ src/station-interface/routes/controls/index.js
@@ -1,0 +1,35 @@
+import Chrony from './chrony.js'
+import Config from './config.js'
+import InternetGateway from './internet-gateway.js'
+import InternetWifiStrength from './internet-wifi-strength.js'
+import ModemDisable from './modem-disable.js'
+import ModemEnable from './modem-enable.js'
+import ModemSignalStrength from './modem-signal-strength.js'
+import ProgramRadios from './program-radios.js'
+import RadioRestart from './radio-restart.js'
+import Reboot from './reboot.js'
+import Software from './software.js'
+import UpdateRebootSchedule from './update-reboot-schedule.js'
+import RebootSchedule from './reboot-schedule.js'
+import Update from './update.js'
+import WifiDisable from './wifi-disable.js'
+import WifiEnable from './wifi-enable.js'
+
+export default {
+  Chrony,
+  Config,
+  InternetGateway,
+  InternetWifiStrength,
+  ModemDisable,
+  ModemEnable,
+  ModemSignalStrength,
+  ProgramRadios,
+  RadioRestart,
+  Reboot,
+  Software,
+  RebootSchedule,
+  UpdateRebootSchedule,
+  Update,
+  WifiDisable,
+  WifiEnable
+}
