+++ src/station-interface/routes/controls/internet-gateway.js
@@ -1,0 +1,11 @@
+export default (req, res) => {
+  fetch('http://localhost:3000/internet/gateway')
+    .then(res => res.json())
+    .then((json) => {
+      res.json(json)
+    })
+    .catch((err) => {
+      console.error(err)
+      res.sendStatus(500)
+    })
+}
