+++ src/station-interface/routes/controls/modem-signal-strength.js
@@ -1,0 +1,11 @@
+export default async (req, res) => {
+  fetch('http://localhost:3000/modem/signal-strength')
+    .then(res => res.json())
+    .then((json) => {
+      res.json(json)
+    })
+    .catch((err) => {
+      console.error(err)
+      res.sendStatus(500)
+    })
+}
