+++ src/station-interface/routes/controls/program-radios.js
@@ -1,0 +1,10 @@
+export default async (req, res) => {
+  fetch('http://localhost:3000/program-radios')
+    .then(() => {
+      res.sendStatus(200)
+    })
+    .catch((err) => {
+      console.error(err)
+      res.sendStatus(500)
+    })
+}
