+++ src/station-interface/routes/controls/wifi-disable.js
@@ -1,0 +1,4 @@
+export default async (req, res) => {
+  await RunCommand('/bin/bash /lib/ctt/sensor-station-software/system/scripts/disable-wifi.sh')
+  return res.status(200).send()
+}
