+++ src/station-interface/routes/crash.js
@@ -1,0 +1,3 @@
+export default Object.freeze((req, res) => {
+  throw new Error('throwing crash error')
+})
