+++ src/station-interface/routes/data/ctt/delete-rotated.js
@@ -1,0 +1,7 @@
+import RunCommand from '../../../../command.js'
+
+export default async (req, res) => {
+  await RunCommand('rm -rf /data/rotated')
+  await RunCommand('mkdir /data/rotated')
+  res.json({ res: true })
+}
