+++ src/station-interface/routes/data/ctt/delete-uploaded.js
@@ -1,0 +1,6 @@
+import RunCommand from '../../../../command.js'
+
+export default async (req, res) => {
+  await RunCommand('rm -rf /data/uploaded/ctt')
+  res.json({ res: true })
+}
