+++ src/station-interface/routes/data/ctt/index.js
@@ -1,0 +1,15 @@
+import DataCurrent from './data-current.js'
+import DataRotated from './data-rotated.js'
+import DataUploaded from './data-uploaded.js'
+import DeleteRotated from './delete-rotated.js'
+import DeleteUploaded from './delete-uploaded.js'
+import LogFile from './log-file.js'
+
+export default {
+  DataCurrent,
+  DataRotated,
+  DataUploaded,
+  DeleteRotated,
+  DeleteUploaded,
+  LogFile,
+}
