+++ src/station-interface/routes/data/index.js
@@ -1,0 +1,9 @@
+import ClearLog from './clear-log.js'
+import Ctt from './ctt/index.js'
+import Sg from './sg/index.js'
+
+export default {
+  ClearLog,
+  Ctt,
+  Sg,
+}
