+++ src/station-interface/routes/data/sg/delete-rotated.js
@@ -1,0 +1,7 @@
+import RunCommand from '../../../../command.js'
+
+export default async (req, res) => {
+  await RunCommand('rm -rf /data/SGdata/*')
+  await RunCommand('systemctl restart sensorgnome')
+  res.json({ res: true })
+}
