+++ src/station-interface/routes/data/sg/delete-uploaded.js
@@ -1,0 +1,6 @@
+import RunCommand from '../../../../command.js'
+
+export default async (req, res) => {
+  await RunCommand('rm -rf /data/uploaded/sg')
+  res.json({ res: true })
+}
