+++ src/station-interface/routes/data/sg/index.js
@@ -1,0 +1,11 @@
+import DeleteRotated from './delete-rotated.js'
+import DeleteUploaded from './delete-uploaded.js'
+import RotatedData from './sg-data-rotated.js'
+import UploadedData from './sg-data-uploaded.js'
+
+export default {
+  DeleteRotated,
+  DeleteUploaded,
+  RotatedData,
+  UploadedData,
+}
