+++ src/station-interface/routes/index.js
@@ -1,653 +1,24 @@
-import express from 'express'
-import { glob } from 'glob'
-import fs from 'fs'
-import moment from 'moment'
-import { spawn } from 'child_process'
-import archiver from 'archiver'
-import bodyParser from 'body-parser'
-import fetch from 'node-fetch'
-import RunCommand from '../../command.js'
-import bcrypt from 'bcrypt'
-import jwt from 'jsonwebtoken'
-import * as crypto from 'crypto'
-import cookieParser from 'cookie-parser'
-import { createDbConnection, getUsers, insertRow, disconnectDb } from '../models/user.js'
+import Main from './main.js'
+import Register from './register.js'
+import Login from './login.js'
+import Blu from './blu.js'
+import Crash from './crash.js'
+import UpdateStation from './update-station.js'
+import Logout from './logout.js'
 
-const TMP_FILE = '/tmp/download.zip'
-const SG_DEPLOYMENT_FILE = '/data/sg_files/deployment.txt'
-const LOG_FILE = '/data/sensor-station.log'
-const ConfigFileURI = '/etc/ctt/station-config.json'
-const secret = crypto.randomBytes(20).toString('hex')
+import Controls from './controls/index.js'
+import Data from './data/index.js'
+import Sensorgnome from './sensorgnome/index.js'
 
-let email, password
-
-const router = express.Router()
-express().use(cookieParser())
-
-// connect to database
-const db = createDbConnection()
-let users = await getUsers(db)
-// disconnect to database
-disconnectDb(db)
-
-const verifySession = async (req, res, next) => {
-  try {
-    const authHeader = req.headers["cookie"]; // get the session cookie from request header
-    // const cookie = authHeader.split("=")[2]; // If there is, split the cookie string to get the actual jwt
-    const cookie = authHeader.match(/(?<=name=).+/g)[0]
-
-    // Verify using jwt to see if token has been tampered with or if it has expired.
-    // that's like checking the integrity of the cookie
-    jwt.verify(cookie, secret, async (err, decoded) => {
-      if (err) {
-        console.log('err', err)
-        // if token has been altered or has expired, return an unauthorized error
-        // return res
-        //   .status(401)
-        //   .json({ message: "This session has expired. Please login" });
-        return res.render('login-fail', {
-          message: 'pug',
-          headerText: 'Session has expired. Please login again.'
-        })
-      }
-
-      const { email } = decoded; // get user id from the decoded token
-      const user = users.find((data) => email === data.email)
-      req.user = user
-    });
-  } catch (err) {
-    res.redirect('/login')
-  }
-}
-
-router.get('/', async function (req, res, next) {
-  // check if user is authenticated
-  if (users.length === 0 && email === undefined && password === undefined) {
-    res.render('main', { title: 'CTT Sensor Station', message: 'pug' })
-
-  } else if (users.length > 0 && email === undefined && password === undefined) {
-    res.redirect('/login')
-
-  } else if (users.length > 0 && email !== undefined && password !== undefined) {
-    await verifySession(req, res)
-
-    if (req.user) {
-      res.render('main', { title: 'CTT Sensor Station', message: 'pug' })
-    }
-
-  } else {
-    res.redirect('/login')
-  }
+export default Object.freeze({
+    Main,
+    Register,
+    Login,
+    Blu,
+    Crash,
+    UpdateStation,
+    Logout,
+    Controls,
+    Data,
+    Sensorgnome,
 })
-
-router.get('/login', (req, res) => {
-  res.render('login', { title: 'CTT Login', message: 'pug' })
-})
-
-router.get('/register', (req, res) => {
-  if (users.length === 0) {
-    res.render('register', { title: 'CTT Registration', message: 'pug' })
-  } else if (email && password && users.length > 0) {
-    res.render('register', { title: 'CTT Registration', message: 'pug' })
-  } else {
-    res.redirect('/login')
-  }
-})
-
-router.post('/register', async (req, res) => {
-  try {
-    if (!req.body.email || !req.body.password) {
-      res.status('400')
-      res.send('Invalid details!')
-    } else {
-      let foundUser = users.find((data) => req.body.email === data.email)
-
-      if (foundUser === undefined) {
-        let hashPassword = await bcrypt.hash(req.body.password, 10)
-
-        // let newUser = {
-        //   id: Date.now(),
-        //   email: req.body.email,
-        //   password: hashPassword,
-        // }
-        const db = createDbConnection()
-        await insertRow(db, req.body.email, hashPassword)
-        users = await getUsers(db)
-        disconnectDb(db)
-
-        res.render('register-success',
-          {
-            title: 'Registration Successful',
-            message: 'pug'
-          })
-      } else if (foundUser) {
-        res.render('register-fail',
-          {
-            headerText: 'Email already exists! Login or choose another email address.',
-            message: 'pug'
-          })
-      } else {
-        res.render('register-fail',
-          {
-            title: 'Registration failed.',
-            headerText: 'An error has occurred. Please try again',
-            message: 'pug'
-          })
-      }
-    }
-  } catch (err) {
-    // res.send('Internal server error', err)
-    res.sendStatus(400).send('Internal server error')
-  }
-})
-
-router.post('/login', async (req, res) => {
-  const db = createDbConnection()
-  let users = await getUsers(db)
-  disconnectDb(db)
-
-  try {
-    let foundUser = users.find((data) => req.body.email === data.email)
-    if (foundUser) {
-      let storedPass = foundUser.password
-      email = foundUser.email
-      password = foundUser.password
-      const passwordMatch = await bcrypt.compare(req.body.password, storedPass)
-
-      if (passwordMatch) {
-        const token = jwt.sign({ email: email },
-          secret,
-          { expiresIn: '5m' })
-        // save session in cookie?
-        res.cookie('name', token, { maxAge: 300000 });
-        res.redirect('/')
-      } else {
-        res.render('login-fail', {
-          message: 'pug',
-          headerText: 'Password is incorrect, please login again.'
-        })
-      }
-
-    } else {
-      let fakePass = `$2b$$10$ifgfgfgfgfgfgfggfgfgfggggfgfgfga`
-      await bcrypt.compare(req.body.password, fakePass)
-      res.render('login-fail', {
-        message: 'pug',
-        headerText: 'Email not found.'
-      })
-    }
-  } catch {
-    res.send('Internal server error')
-  }
-})
-
-router.get('/login-fail', async (req, res) => {
-  res.render('login-fail', { title: 'Login Failed', message: 'pug' })
-})
-
-router.post('/register-success', async (req, res) => {
-  try {
-    let foundUser = users.find((data) => req.body.email === data.email)
-    if (foundUser) {
-      let submittedPass = req.body.password
-      let storedPass = foundUser.password
-
-      email = foundUser.email
-      password = foundUser.password
-      const passwordMatch = await bcrypt.compare(submittedPass, storedPass)
-
-      if (passwordMatch) {
-        res.render('main', { title: 'CTT Sensor Station', message: 'pug' })
-
-      } else {
-        res.redirect('/login-fail')
-
-        // res.send("<div align='center'><h2>Invalid email or password</h2></div><br><br><div align='center'><a href='./login.pug'>login again</a></div>")
-      }
-    } else {
-      let fakePass = `$2b$$10$ifgfgfgfgfgfgfggfgfgfggggfgfgfga`
-      await bcrypt.compare(req.body.password, fakePass)
-
-      res.send("<div align='center'><h2>Invalid email or password</h2></div><br><br><div align='center'><a href='./login.pug'>login again<a><div>")
-    }
-  } catch {
-    res.send('Internal server error')
-  }
-})
-
-router.get('/logout', async (req, res) => {
-  res.clearCookie('name')
-  res.redirect('/login')
-})
-
-router.get('/blu', function (req, res, next) {
-  // check if user is authenticated
-  // console.log('request', req.session)
-  if (users.length === 0) {
-    res.render('main-blu', { title: 'CTT Blu Receiver Interface', message: 'pug' })
-  } else if (users.length > 0 && email && password) {
-    res.render('login', { title: 'CTT Login', message: 'pug' })
-
-  } else {
-    res.redirect('/login')
-  }
-})
-
-router.get('/update-station', function (req, res, next) {
-  res.render('station-update')
-})
-
-router.get('/crash', (req, res, next) => {
-  // crash - bad variable
-  throw (Error('throwing crash error'))
-})
-
-router.get('/sg-deployment', (req, res, next) => {
-  fs.readFile(SG_DEPLOYMENT_FILE, (err, contents) => {
-    if (err) {
-      next(err)
-    } else {
-      res.send(contents)
-    }
-  })
-})
-
-router.post('/save-sg-deployment', (req, res, next) => {
-  fs.writeFile(SG_DEPLOYMENT_FILE, req.body.contents, (err) => {
-    if (err) {
-      next(err)
-    } else {
-      console.log('saved data')
-      res.json({ res: true })
-    }
-  })
-})
-
-const BASE_SG_TAG_DB_NAME = 'SG_tag_database'
-router.post('/upload-sg-tag-file', (req, res) => {
-  console.log('tag database upload')
-  const ext = req.get('file-extension')
-  const filename = `${BASE_SG_TAG_DB_NAME}.${ext}`
-  console.log('about to delete sg tag db files')
-  RunCommand(`rm /data/sg_files/${BASE_SG_TAG_DB_NAME}*`)
-    .then(() => {
-      let uri = `/data/sg_files/${filename}`
-      console.log('writing tag database file')
-      fs.writeFileSync(uri, req.body)
-      res.json({ res: true })
-    }).catch((err) => {
-      console.log('something went wrong handling new SG tag database file')
-      console.error(err)
-    })
-})
-
-const prepareData = (filelist) => {
-  return new Promise((resolve, reject) => {
-    if (fs.existsSync(TMP_FILE)) {
-      fs.unlinkSync(TMP_FILE)
-    }
-    let output = fs.createWriteStream(TMP_FILE)
-    output.on('close', () => {
-      resolve(true)
-    })
-    output.on('error', (err) => {
-      reject(err)
-    })
-    var archive = archiver('zip', {
-      zlip: { level: 9 }
-    })
-    archive.on('error', (err) => {
-      reject(err)
-    })
-    archive.pipe(output)
-    filelist.forEach((filename) => {
-      archive.file(filename, { name: filename })
-    })
-    archive.finalize()
-  })
-}
-
-router.get('/ctt-data-current', async (req, res, next) => {
-  // glob('/data/*.csv', (err, filelist) => {
-  let filelist = await glob('/data/*.csv')
-  if (filelist.length < 1) {
-    res.send('No data available')
-    return
-  }
-  prepareData(filelist).then((prepare_result) => {
-    let download_name = `ctt-data.${moment(new Date()).format('YYYY-MM-DD_HHMMSS')}.zip`
-    res.download(TMP_FILE, download_name)
-  }).catch((err) => {
-    next(err)
-  })
-  // })
-})
-
-router.get('/ctt-logfile', async (req, res, next) => {
-  let filelist = await glob('/data/CTT-*-log.csv')
-  console.log('ctt logfile await', filelist)
-  // glob('/data/CTT-*-log.csv', (err, filelist) => {
-  if (filelist.length < 1) {
-    res.send('no log file to send')
-    return
-  }
-  prepareData(filelist).then((prepare_result) => {
-    let download_name = `ctt-log.${moment(new Date()).format('YYYY-MM-DD_HHMMSS')}.zip`
-    res.download(TMP_FILE, download_name)
-  }).catch((err) => {
-    res.send('error preparing ctt log files ' + err)
-  })
-  // })
-})
-
-router.get('/sg-data-rotated', async (req, res, next) => {
-  let filelist = await glob('/data/SGdata/*/*.gz')
-  // glob('/data/SGdata/*/*.gz', (err, filelist) => {
-  if (filelist.length < 1) {
-    res.send('Nothing to download yet')
-    return
-  }
-  prepareData(filelist).then((prepare_result) => {
-    let download_name = `sg-data-uploaded.${moment(new Date()).format('YYYY-MM-DD_HHMMSS')}.zip`
-    res.download(TMP_FILE, download_name)
-  }).catch((err) => {
-    res.send('ERROR processing download request ' + err)
-  })
-  // })
-})
-
-router.get('/sg-data-uploaded', async (req, res, next) => {
-  let filelist = await glob('/data/uploaded/sg/*/*.gz')
-  // glob('/data/uploaded/sg/*/*.gz', (err, filelist) => {
-  if (filelist.length < 1) {
-    res.send('Nothing to download yet')
-    return
-  }
-  prepareData(filelist).then((prepare_result) => {
-    let download_name = `sg-data-uploaded.${moment(new Date()).format('YYYY-MM-DD_HHMMSS')}.zip`
-    res.download(TMP_FILE, download_name)
-  }).catch((err) => {
-    res.send('ERROR processing download request ' + err)
-  })
-  // })
-})
-
-router.get('/ctt-data-rotated', async (req, res, next) => {
-  // glob('/data/rotated/*.gz', (err, filelist) => {
-  let filelist = await glob('/data/rotated/*.gz')
-  if (filelist.length < 1) {
-    res.send('Nothing to download yet')
-    return
-  }
-  prepareData(filelist).then((prepare_result) => {
-    let download_name = `ctt-data-rotated.${moment(new Date()).format('YYYY-MM-DD_HHMMSS')}.zip`
-    res.download(TMP_FILE, download_name)
-  }).catch((err) => {
-    res.send('ERROR processing download request ' + err)
-  })
-  // })
-})
-
-router.get('/ctt-data-uploaded', async (req, res, next) => {
-  // glob('/data/uploaded/ctt/*/*.gz', (err, filelist) => {
-  let filelist = await glob('/data/uploaded/ctt/*/*.gz')
-  if (filelist.length < 1) {
-    res.send('Nothing to download yet')
-    return
-  }
-  prepareData(filelist).then((prepare_result) => {
-    let download_name = `ctt-data-uploaded.${moment(new Date()).format('YYYY-MM-DD_HHMMSS')}.zip`
-    res.download(TMP_FILE, download_name)
-  }).catch((err) => {
-    res.send('ERROR processing download request ' + err)
-  })
-  // })
-})
-
-router.post('/delete-ctt-data-uploaded', async (req, res) => {
-  await RunCommand('rm -rf /data/uploaded/ctt')
-  res.json({ res: true })
-})
-
-router.post('/delete-ctt-data-rotated', async (req, res) => {
-  await RunCommand('rm -rf /data/rotated')
-  await RunCommand('mkdir /data/rotated')
-  res.json({ res: true })
-})
-
-router.post('/delete-sg-data-uploaded', async (req, res) => {
-  await RunCommand('rm -rf /data/uploaded/sg')
-  res.json({ res: true })
-})
-
-router.post('/delete-sg-data-rotated', async (req, res) => {
-  await RunCommand('rm -rf /data/SGdata/*')
-  await RunCommand('systemctl restart sensorgnome')
-  res.json({ res: true })
-})
-
-router.post('/clear-log/', async (req, res, next) => {
-  // let log_file = '/data/sensor-station.log'
-  let log_file = await glob('/data/CTT-*-log.csv')
-  console.log('log file', log_file)
-  if (fs.existsSync(log_file[0])) {
-    fs.unlinkSync(log_file[0])
-    res.send(JSON.stringify({ res: true }))
-    return
-  }
-  res.send(JSON.stringify({ res: false }))
-})
-
-router.get('/chrony', (req, res, next) => {
-  const cmd = spawn('chronyc', ['sources', '-v'])
-  let buffer = ''
-  cmd.stdout.on('data', (data) => {
-    buffer += data.toString()
-  })
-  cmd.on('close', () => {
-    res.send(buffer)
-  })
-})
-
-router.post('/reboot', (req, res, next) => {
-  const reboot = spawn('shutdown', ['-r', 'now'])
-  reboot.stdout.on('data', (data) => {
-    console.log('data', data.toString())
-  })
-  reboot.stderr.on('data', (data) => {
-    console.log('err', data.toString())
-  })
-  res.send('rebooting')
-})
-
-router.get('/update', (req, res, next) => {
-  res.render('update', { title: 'CTT Sensor Station Update', message: 'pug' })
-})
-
-router.use(bodyParser.raw({
-  limit: '50mb'
-}))
-
-router.post('/update', (req, res, next) => {
-  let contents = req.body.toString()
-  fs.writeFile('/tmp/update.sh', contents, (err) => {
-    if (err) {
-      res.send('error writing update file')
-      return
-    } else {
-      const cmd = spawn('/bin/bash', ['/tmp/update.sh'])
-      cmd.on('error', (err) => {
-        console.error(err)
-      })
-      cmd.stdout.on('data', (data) => {
-        console.log(data)
-      })
-      cmd.stderr.on('data', (data) => {
-        console.log('error', data)
-      })
-
-      cmd.on('close', () => {
-        res.send('updating')
-      })
-    }
-  })
-})
-
-router.get('/config', (req, res, next) => {
-  try {
-    let config = JSON.parse(fs.readFileSync(ConfigFileURI).toString())
-    res.json(config)
-  } catch (err) {
-    res.json({ err: err.toString() })
-  }
-})
-
-router.post('/radio-restart', (req, res) => {
-  const cmd = spawn('systemctl', ['restart', 'station-radio-interface'])
-  console.log('issuing radio restart')
-  cmd.on('error', (err) => {
-    console.error(err)
-    res.sendStatus(500)
-  })
-  cmd.stdout.on('data', (data) => {
-    console.log(data.toString())
-  })
-  cmd.on('close', () => {
-    res.sendStatus(204)
-  })
-})
-
-router.post('/program-radios', async (req, res) => {
-  fetch('http://localhost:3000/program-radios')
-    .then(() => {
-      res.sendStatus(200)
-    })
-    .catch((err) => {
-      console.error(err)
-      res.sendStatus(500)
-    })
-})
-
-/**
- * 
- * get software versions from package.json data
- */
-router.get('/software', (req, res) => {
-  fetch('http://localhost:3000/node/version')
-    .then(res => res.json())
-    .then((json) => {
-      res.json(json)
-    })
-    .catch((err) => {
-      console.error(err)
-      res.sendStatus(500)
-    })
-})
-
-/**
- * proxy to hardware server to get internet gateway
- */
-router.get('/internet-gateway', (req, res) => {
-  fetch('http://localhost:3000/internet/gateway')
-    .then(res => res.json())
-    .then((json) => {
-      res.json(json)
-    })
-    .catch((err) => {
-      console.error(err)
-      res.sendStatus(500)
-    })
-})
-
-/**
- * proxy to hardware server to get internet gateway
- */
-router.get('/internet-wifi-strength', (req, res) => {
-  // fetch('http://localhost:3000/internet/wifi-strength')
-  fetch('http://localhost:3000/internet/wifi-networks')
-
-    .then(res => res.json())
-    .then((json) => {
-      res.json(json)
-    })
-    .catch((err) => {
-      console.error(err)
-      res.sendStatus(500)
-    })
-})
-
-/**
- * get reboot schedule from crontab via hardware server proxy
- */
-router.get('/reboot-schedule', (req, res) => {
-  fetch('http://localhost:3000/control/reboot-schedule')
-    .then(res => res.json())
-    .then((json) => {
-      res.json(json)
-    })
-    .catch((err) => {
-      console.error(err)
-      res.sendStatus(500)
-    })
-})
-
-/**
- * update reboot scheule via hardware server proxy
- */
-router.post('/update-reboot-schedule', (req, res) => {
-  console.log('rx', req.body)
-  fetch('http://localhost:3000/control/update-reboot-schedule', {
-    method: 'post',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(req.body)
-  })
-    .then((proxy_response) => {
-      if (proxy_response.ok) {
-        res.sendStatus(204)
-      } else {
-        res.sendStatus(proxy_response.status)
-      }
-    })
-    .catch((err) => {
-      console.error('error with reboot schedule  proxy post')
-      console.error(err)
-      res.sendStatus(500)
-    })
-})
-
-router.post('/modem/enable', async (req, res) => {
-  // await RunCommand('nmcli connection up station-modem')
-  await RunCommand('/bin/bash /lib/ctt/sensor-station-software/system/scripts/enable-modem.sh')
-
-  return res.status(200).send()
-})
-router.post('/modem/disable', async (req, res) => {
-  // await RunCommand('nmcli connection down station-modem')
-  await RunCommand('/bin/bash /lib/ctt/sensor-station-software/system/scripts/disable-modem.sh')
-
-  return res.status(200).send()
-})
-
-router.get('/modem-signal-strength', async (req, res) => {
-  fetch('http://localhost:3000/modem/signal-strength')
-    .then(res => res.json())
-    .then((json) => {
-      res.json(json)
-    })
-    .catch((err) => {
-      console.error(err)
-      res.sendStatus(500)
-    })
-})
-
-router.post('/wifi/enable', async (req, res) => {
-  await RunCommand('/bin/bash /lib/ctt/sensor-station-software/system/scripts/enable-wifi.sh')
-  return res.status(200).send()
-})
-
-router.post('/wifi/disable', async (req, res) => {
-  await RunCommand('/bin/bash /lib/ctt/sensor-station-software/system/scripts/disable-wifi.sh')
-  return res.status(200).send()
-})
-
-export default router
