+++ src/station-interface/routes/logout.js
@@ -1,0 +1,6 @@
+import UserApi from '../users.js'
+
+export default Object.freeze((req, res) => {
+  res.clearCookie(UserApi.CookieName)
+  res.redirect('/login')
+})
