+++ src/station-interface/routes/main.js
@@ -1,0 +1,3 @@
+export default (req, res) => {
+  res.render('main', { title: 'CTT Sensor Station', message: 'pug' })
+}
