+++ src/station-interface/routes/sensorgnome/files.js
@@ -1,0 +1,3 @@
+export default Object.freeze({
+    Deployment: '/data/sg_files/deployment.txt',
+})
