+++ src/station-interface/routes/sensorgnome/index.js
@@ -1,0 +1,9 @@
+import SaveSgDeployment from './save-sg-deployment.js'
+import Sgdeployment from './sg-deployment.js'
+import UploadSgTagFile from './upload-sg-tag-file.js'
+
+export default {
+  SaveSgDeployment,
+  Sgdeployment,
+  UploadSgTagFile,
+}
