+++ src/station-interface/routes/sensorgnome/save-sg-deployment.js
@@ -1,0 +1,13 @@
+import fs from 'fs'
+import SensorgnomeFiles from './files.js'
+
+export default (req, res, next) => {
+  fs.writeFile(SensorgnomeFiles.Deployment, req.body.contents, (err) => {
+    if (err) {
+      next(err)
+    } else {
+      console.log('saved data')
+      res.json({ res: true })
+    }
+  })
+}
