+++ src/station-interface/routes/update-station.js
@@ -1,0 +1,3 @@
+export default Object.freeze((req, res) => {
+  res.render('station-update')
+})
